+++ include/kansei/fallen/determinant/determinant_type.hpp
@@ -18,8 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#ifndef KANSEI__DETERMINANT__DETERMINANT_TYPE_HPP_
-#define KANSEI__DETERMINANT__DETERMINANT_TYPE_HPP_
+#ifndef KANSEI__FALLEN__DETERMINANT__DETERMINANT_TYPE_HPP_
+#define KANSEI__FALLEN__DETERMINANT__DETERMINANT_TYPE_HPP_
 
 namespace kansei
 {
@@ -32,4 +32,4 @@ enum DeterminantType
 
 }  // namespace kansei
 
-#endif  // KANSEI__DETERMINANT__DETERMINANT_TYPE_HPP_
+#endif  // KANSEI__FALLEN__DETERMINANT__DETERMINANT_TYPE_HPP_
