+++ include/kansei/measurement/filter/madgwick/world_frame.hpp
@@ -18,8 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#ifndef KANSEI__UNIT__FILTER__MADGWICK__WORLD_FRAME_HPP_
-#define KANSEI__UNIT__FILTER__MADGWICK__WORLD_FRAME_HPP_
+#ifndef KANSEI__MEASUREMENT__FILTER__MADGWICK__WORLD_FRAME_HPP_
+#define KANSEI__MEASUREMENT__FILTER__MADGWICK__WORLD_FRAME_HPP_
 
 namespace kansei
 {
@@ -33,4 +33,4 @@ enum WorldFrame
 
 }  // namespace kansei
 
-#endif  // KANSEI__UNIT__FILTER__MADGWICK__WORLD_FRAME_HPP_
+#endif  // KANSEI__MEASUREMENT__FILTER__MADGWICK__WORLD_FRAME_HPP_
